casting to int

+++ PulseGeneration.py
@@ -311,7 +311,7 @@ def anti_plume_pulse(sampling_rate, params):
     plume = plume['plume'][0]
 
     # resample to match sampling rate
-    resampled = signal.resample(plume, len(plume)*(sampling_rate / params['data_fs']))
+    resampled = signal.resample(plume, int(len(plume)*(sampling_rate / params['data_fs'])))
     # zero out negative values
     resampled[resampled < 0] = 0
     # normalise
